Migrate Query.hasCollectedOperations
